Sheet: Reorganize and rename methods. PlanCostCalculator: implement a deferred version (PlanCostCalculatorDeferred) by means of the Sheet algebra. SBT: increase version 1.7.1 to 1.8.2

+++ src/main/scala/wms/flow/planner/engine/RequiredPowerCalculator.scala
@@ -99,11 +99,11 @@ class RequiredPowerCalculator[PA <: PiecewiseAlgebra, CG <: ClosedGraph](val pie
 
 						val desiredLoadAtEndingInstantAtStageAtPiece: Quantity = desiredBacklogAtEndingInstantAtStage.getPieceMeanAt(pieceIndex) match {
 							case Maximal =>
-								backlogCapacity(stage);
+								backlogCapacity(stage).toFloat;
 
 							case Minimal(duration) =>
 								scala.math.min(
-									backlogCapacity(stage),
+									backlogCapacity(stage).toFloat,
 									trajectoryOfLoadDemandedByDownstream.integrate(end, end + duration, true)
 								);
 						}
